Define a set of objects by filtering according to some criteria

Define a set of objects by filtering according to some criteria

api_define_object_set(
  project_id,
  taxon_ids = NULL,
  include_children = NULL,
  status = NULL,
  sample_ids = NULL,
  depth_below = NULL,
  depth_above = NULL,
  lat_min = NULL,
  lat_max = NULL,
  lon_min = NULL,
  lon_max = NULL,
  from_date = NULL,
  to_date = NULL,
  months = NULL,
  from_time = NULL,
  to_time = NULL,
  invert_time_span = NULL,
  time_of_day = NULL,
  instrument = NULL,
  annotator_ids = NULL,
  annotated_after = NULL,
  annotated_before = NULL,
  custom_numeric_field = NULL,
  custom_numeric_min = NULL,
  custom_numeric_max = NULL,
  custom_text_field = NULL,
  custom_text_contains = NULL
)

Arguments

project_id
numeric id of the project to filter in (a single one)

taxon_ids
numeric ids of the taxa to filter

include_children
boolean, whether include the taxa's children

status
one of "U":unclassified, "P":predicted, "NV":not validated, "V":validated, "PV":predicted or validated, "NVM":validated by others, "VM":validated by me, "D":dubious

sample_ids
numeric ids of the samples to filter

depth_below, depth_above
range of depths to filter; the bounds of the range are included (i.e. depth_below=10 keeps objects situated at 10m depth). TODO more info about intersection with depth_min and max of objects

lat_min, lat_max
range of latitudes to filter

lon_min, lon_max
range of longitudes to filter

from_date, to_date
range of dates to filter; should be of class "Date"

months
vector of months numbers to filter (from 1 to 12)

from_time, to_time
range of times to filter; should be a character, coercible to a time within a day (e.g. "13:00:00" for 13:00 = 1PM)

invert_time_span
filter the

time_of_day
.

instrument
.

annotator_ids
.

annotated_after
.

annotated_before
.

custom_numeric_field
.

custom_numeric_min
.

custom_numeric_max
.

custom_text_field
.

custom_text_contains
.

Value

A list defining an object set

Examples

# o <- api_define_object_set(185, status="V")