Batch update metadata of objects

Batch update metadata of objects

api_update_object_set(object_ids, updates)

Arguments

object_ids
internal, numeric ids of the objects.

updates
named list with the metadata field(s) to update and the updated value(s). Use api_object() to find their names. Due to a current bug, the list should contain either regular fields only or free fields only, but not a mix of both.

Value

The number of updated objects.

Examples

# get list of possible fields via the `object` endpoint
str(api_object(24473014))

#> List of 28
#>  $ orig_id                   : chr "wp220160223_d1_1_693"
#>  $ object_link               : NULL
#>  $ objid                     : int 24473014
#>  $ projid                    : int 185
#>  $ objdate                   : chr "2016-02-23"
#>  $ objtime                   : chr "10:00:00"
#>  $ latitude                  : num 43.7
#>  $ longitude                 : num 7.32
#>  $ depth_min                 : num 0
#>  $ depth_max                 : num 75
#>  $ sunpos                    : chr "D"
#>  $ classif_id                : int 11509
#>  $ classif_qual              : chr "P"
#>  $ classif_who               : int 15
#>  $ classif_when              : chr "2016-05-26T16:37:20.168202"
#>  $ classif_auto_id           : int 11509
#>  $ classif_auto_score        : num 0.748
#>  $ classif_auto_when         : chr "2020-12-11T19:30:11.949883"
#>  $ classif_crossvalidation_id: NULL
#>  $ img0id                    : int 24467468
#>  $ imgcount                  : int 1
#>  $ complement_info           : NULL
#>  $ similarity                : NULL
#>  $ random_value              : int 58488508
#>  $ sampleid                  : int 15709
#>  $ acquisid                  : int 204
#>  $ images                    :'data.frame':	1 obs. of  10 variables:
#>   ..$ imgid          : int 24467468
#>   ..$ objid          : int 24473014
#>   ..$ imgrank        : int 0
#>   ..$ file_name      : chr "2446/7468.jpg"
#>   ..$ orig_file_name : chr "wp220160223_d1_1_693.jpg"
#>   ..$ width          : int 114
#>   ..$ height         : int 151
#>   ..$ thumb_file_name: logi NA
#>   ..$ thumb_width    : logi NA
#>   ..$ thumb_height   : logi NA
#>  $ free_columns              :List of 69
#>   ..$ lat_end     : num 9999
#>   ..$ lon_end     : num -9999
#>   ..$ area        : num 1028
#>   ..$ mean        : num 224
#>   ..$ stddev      : num 16.6
#>   ..$ mode        : num 241
#>   ..$ min         : num 162
#>   ..$ max         : num 243
#>   ..$ x           : num 51.4
#>   ..$ y           : num 61.5
#>   ..$ xm          : num 51.2
#>   ..$ ym          : num 61.1
#>   ..$ perim.      : num 421
#>   ..$ bx          : num 14626
#>   ..$ by          : num 13245
#>   ..$ width       : num 70
#>   ..$ height      : num 86
#>   ..$ major       : num 51
#>   ..$ minor       : num 25.6
#>   ..$ angle       : num 63.2
#>   ..$ circ.       : num 0.073
#>   ..$ feret       : num 105
#>   ..$ intden      : num 230234
#>   ..$ median      : num 229
#>   ..$ skew        : num -1.14
#>   ..$ kurt        : num 0.603
#>   ..$ %area       : num 0
#>   ..$ xstart      : num 14688
#>   ..$ ystart      : num 13245
#>   ..$ area_exc    : num 1028
#>   ..$ fractal     : num 1.19
#>   ..$ skelarea    : num 199
#>   ..$ slope       : num 0.128
#>   ..$ histcum1    : num 215
#>   ..$ histcum2    : num 228
#>   ..$ histcum3    : num 235
#>   ..$ xmg5        : num 0
#>   ..$ ymg5        : num 0
#>   ..$ nb1         : num 1
#>   ..$ nb2         : num 1
#>   ..$ nb3         : num 0
#>   ..$ compentropy : num 0
#>   ..$ compmean    : num 0
#>   ..$ compslope   : num 0
#>   ..$ compm1      : num 0
#>   ..$ compm2      : num 0
#>   ..$ compm3      : num 0
#>   ..$ symetrieh   : num 7.73
#>   ..$ symetriev   : num 7.96
#>   ..$ symetriehc  : num 7
#>   ..$ symetrievc  : num 7
#>   ..$ convperim   : num 307
#>   ..$ convarea    : num 3196
#>   ..$ fcons       : num 5.57
#>   ..$ thickr      : num 3.11
#>   ..$ tag         : num 1
#>   ..$ esd         : num 654
#>   ..$ elongation  : num 1.96
#>   ..$ range       : num 81
#>   ..$ meanpos     : num -0.306
#>   ..$ centroids   : num -1
#>   ..$ cv          : num 7.59
#>   ..$ sr          : num 21
#>   ..$ perimareaexc: num 0.41
#>   ..$ feretareaexc: num 0.102
#>   ..$ perimferet  : num 4.01
#>   ..$ perimmajor  : num 8.25
#>   ..$ circex      : num 30.5
#>   ..$ cdexc       : num -0.000973

# change one field of metadata in one object
api_update_object_set(24473014, list(longitude=10))

#> [1] 1

# check that the change is effective
api_object(24473014)$longitude

#> [1] 10

# change several fields in one object
api_object(24473014)[c("longitude", "latitude")]

#> $longitude
#> [1] 10
#> 
#> $latitude
#> [1] 43.685
#> 

api_update_object_set(24473014, list(longitude=0, latitude=0))

#> [1] 1

api_object(24473014)[c("longitude", "latitude")]

#> $longitude
#> [1] 0
#> 
#> $latitude
#> [1] 0
#> 

api_object(24473014)$free_columns[c("area", "mean")]

#> $area
#> [1] 1028
#> 
#> $mean
#> [1] 224
#> 

api_update_object_set(24473014, list(area=1, mean=1))

#> [1] 1

api_object(24473014)$free_columns[c("area", "mean")]

#> $area
#> [1] 1
#> 
#> $mean
#> [1] 1
#> 

# change several fields in several objects
api_update_object_set(c(24473014, 24473015), list(longitude=5, latitude=-2))

#> [1] 2

api_object(24473014)[c("longitude", "latitude")]

#> $longitude
#> [1] 5
#> 
#> $latitude
#> [1] -2
#> 

api_object(24473015)[c("longitude", "latitude")]

#> $longitude
#> [1] 5
#> 
#> $latitude
#> [1] -2
#> 

# change the taxonomic identification
api_update_object_set(24473014, list(classif_id=1, classif_qual="V"))

#> [1] 1

api_object(24473014)[c("classif_id", "classif_qual")]

#> $classif_id
#> [1] 1
#> 
#> $classif_qual
#> [1] "V"
#> 

# restore initial values
api_update_object_set(
  c(24473014, 24473015),
  list(longitude=7.31567, latitude=43.68500)
) # NB: these objects were actually collected in the same location

#> [1] 2

api_update_object_set(24473014, list(classif_id=11509, classif_qual="P"))

#> [1] 1

api_update_object_set(24473014, list(area=1028, mean=224))

#> [1] 1

api_update_object_set(24473015, list(area=1118, mean=225))

#> [1] 1