Get information about a single user, by id

Get information about a single user, by id

api_user(user_id)

Arguments

user_id
internal, numeric id of the user.

Value

A list of users or a single user, each containing

id: the unique numeric id of this user;

email: email address, as text;

name: name, as text;

organisation: organisation name, as text;

active: whether the user is still active;

country: the country name, as text (but chosen in a consistent list);

usercreationdate: the date of creation of the user, as text formatted according to the ISO 8601 standard; parse it with lubridate::ymd_hms();

usercreationreason: paragraph describing the usage of EcoTaxa made by the user.

See also

Examples

api_user(993)

#> $id
#> [1] 993
#> 
#> $email
#> [1] "ecotaxa.api.user@gmail.com"
#> 
#> $name
#> [1] "Test API user"
#> 
#> $organisation
#> [1] "Oceanographic Laboratory of Villefranche sur Mer - LOV"
#> 
#> $active
#> [1] TRUE
#> 
#> $country
#> [1] "France"
#> 
#> $usercreationdate
#> [1] "2020-11-05T12:31:48.299713"
#> 
#> $usercreationreason
#> [1] "Test API"
#>