Search for users, by name

Search for users, by name

api_users_search(name)

Arguments

name
string to look for in the user name

Value

A list of users or a single user, each containing

id: the unique numeric id of this user;

email: email address, as text;

name: name, as text;

organisation: organisation name, as text;

active: whether the user is still active;

country: the country name, as text (but chosen in a consistent list);

usercreationdate: the date of creation of the user, as text formatted according to the ISO 8601 standard; parse it with lubridate::ymd_hms();

usercreationreason: paragraph describing the usage of EcoTaxa made by the user.

See also

Examples

api_users_search("api")

#>    id                          email            name
#> 1 112 kyle.capistrantfossa@maine.edu  Kyle Capistran
#> 2 632           k.ferrapie@monego.mc Karine Ferrapie
#> 3 993     ecotaxa.api.user@gmail.com   Test API user
#>                                             organisation active       country
#> 1                              Maine University - Umaine   TRUE United States
#> 2                            Société Monegasque des Eaux   TRUE        France
#> 3 Oceanographic Laboratory of Villefranche sur Mer - LOV   TRUE        France
#>             usercreationdate usercreationreason
#> 1                       <NA>               <NA>
#> 2 2019-05-23T15:20:07.647839                   
#> 3 2020-11-05T12:31:48.299713           Test API